Created SystematicModelOptimization
The new file SystematicModelOptimization systematically tests number of iterations and model architectures for the dataset. It then reports the test metrics, the optimal parameters and re-trains the best model on the whole dataset. Finally it saves the model and a txt file with the relevant metrics and parameters to recreate the model.